2.0.0 put declaration and diagnostic in its own modules
